test: gvb_interp

+++ gvb_interp/src/ast.rs
@@ -128,7 +128,7 @@ impl Range {
   }
 
   pub fn range(&self) -> std::ops::Range<usize> {
-    (self.start as usize)..(self.end as usize)
+    self.start..self.end
   }
 }
